fix: rustfmt apply.rs and symbol_index.rs for CI (#110)
* style: rustfmt apply.rs and symbol_index.rs for CI

* fix: remove unused has_mod import in expect_actual test

* fix: remove unnecessary as u32 cast in expect_actual test

+++ src/cli/expect_actual_tests.rs
@@ -1,4 +1,4 @@
-use super::{detect_ss, first_id, has_mod, parse_kotlin};
+use super::{detect_ss, first_id, parse_kotlin};
 use std::path::Path;
 
 #[test]
@@ -24,7 +24,7 @@ fn first_id_no_panic() {
     let src = "class Foo";
     let tree = parse_kotlin(src).unwrap();
     let root = tree.root_node();
-    if let Some(decl) = root.child(0 as u32) {
+    if let Some(decl) = root.child(0) {
         let _id = first_id(&decl, src);
     }
 }
